requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
black==23.12.1
boto3==1.40.21
botocore==1.40.21
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.10.6
crypto==1.4.1
cryptography==45.0.7
Django==4.2.23
django-anymail==10.3
django-axes==6.5.2
django-braces==1.15.0
django-cors-headers==4.7.0
django-debug-toolbar==4.4.6
django-decouple==2.1
django-environ==0.12.0
django-filter==23.5
django-ratelimit==4.1.0
django-redis==5.4.0
django-split-settings==1.0.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
factory_boy==3.3.3
Faker==19.13.0
flake8==6.1.0
flower==2.0.1
gunicorn==21.2.0
humanize==4.13.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==5.13.2
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kombu==5.5.4
mccabe==0.7.0
mypy_extensions==1.1.0
Naked==0.1.32
packaging==25.0
pathspec==0.12.1
paystackapi==2.1.3
pillow==10.4.0
platformdirs==4.4.0
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycodestyle==2.11.1
pycparser==2.22
pycryptodome==3.17
pyflakes==3.1.0
PyJWT==2.10.1
pytest==7.4.4
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
s3transfer==0.13.1
sentry-sdk==1.45.1
shellescape==3.8.1
six==1.17.0
sqlparse==0.5.3
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
